Move import_module shim to utils for compat across the codebase
