+++ lib/ansible/executor/powershell/module_manifest.py
@@ -14,16 +14,11 @@ import re
 
 from distutils.version import LooseVersion
 
-# HACK: keep Python 2.6 controller tests happy in CI until they're properly split
-try:
-    from importlib import import_module
-except ImportError:
-    import_module = __import__
-
 from ansible import constants as C
 from ansible.errors import AnsibleError
 from ansible.module_utils._text import to_bytes, to_native, to_text
 from ansible.plugins.loader import ps_module_utils_loader
+from ansible.utils.import_module import import_module
 
 
 class PSModuleDepFinder(object):
